% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{$.re2_matrix}
\alias{$.re2_matrix}
\title{Get a match group}
\usage{
\method{$}{re2_matrix}(self, group)
}
\arguments{
\item{self}{re2_matrix}

\item{group}{group name}
}
\description{
Get a match group
}
\examples{
text = c("this is test",
    "this is test, and this is not test",
    "they are tests")
res = re2_match(
    string = text,
    pattern = "(?P<testname>this)( is)"
)
class(res)
is.matrix(res)
is.character(res)
print(res)
res$testname
res$.match
res$`.2`
res[, ".2"]
res[, ".match"]
res[, "testname"]
}
