% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{get_expression_size}
\alias{get_expression_size}
\title{Get pre-compiled regular expression program size}
\usage{
get_expression_size(pattern, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{...}{further arguments passed to \code{\link{re2}}}
}
\value{
a integer
}
\description{
Returns the program size, a very approximate measure of a regexp's "cost".
Larger numbers are more expensive than smaller numbers.
}
\examples{
get_expression_size(re2("1"))
get_expression_size(re2("(1)"))
get_expression_size(re2("(?:(?:(?:(?:(?:.)?){100})*)+)"))
}
