% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_named_groups}
\alias{get_named_groups}
\title{Return capturing names with indices.}
\usage{
get_named_groups(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
capturing names with indices.
}
\description{
The map records the index of the leftmost group with the given name.
}
\examples{
regexp = re2("(?P<A>expr(?P<B>expr)(?P<C>expr))((expr)(?P<D>expr))")
(res = get_named_groups(regexp))
names(res)
}

