% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_namedcapturinggroups}
\alias{get_namedcapturinggroups}
\title{Return capturing names with indices.}
\usage{
get_namedcapturinggroups(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
capturing names with indices.
}
\description{
Return capturing names with indices.
}
\examples{
regexp = re2_compile("1")
get_pattern(regexp)
}

