% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_numberofcapturinggroups}
\alias{get_numberofcapturinggroups}
\title{Return the number of capturing subpatterns}
\usage{
get_numberofcapturinggroups(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
a integer
}
\description{
Return the number of capturing subpatterns, or -1 if the
regexp wasn't valid on construction.  The overall match ($0)
does not count: if the regexp is "(a)(b)", returns 2.
}
\examples{
regexp = re2_compile("1")
get_numberofcapturinggroups(regexp)
}

