% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_options}
\alias{get_options}
\title{Return the options for a pre-compiled regular expression}
\usage{
get_options(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
a list
}
\description{
Return the options for a pre-compiled regular expression
}
\examples{
get_options(re2("1"))
}

