% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{get_pattern}
\alias{get_pattern}
\title{The string specification for this RE2.}
\usage{
get_pattern(pattern, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{...}{further arguments passed to \code{\link{re2}}}
}
\value{
a string
}
\description{
The string specification for this RE2.
}
\examples{
regexp = re2("1")
get_pattern(regexp)

get_pattern(re2("^(?P<abc>abc)a"))
}

