% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanout.R
\name{get_program_fanout}
\alias{get_program_fanout}
\title{Get program fanout
Outputs the program fanout as a histogram bucketed by powers of 2.
Returns the number of the largest non-empty bucket.}
\usage{
get_program_fanout(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\description{
Get program fanout
Outputs the program fanout as a histogram bucketed by powers of 2.
Returns the number of the largest non-empty bucket.
}

