% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_programsize}
\alias{get_programsize}
\title{Get pre-compiled regular expression program size}
\usage{
get_programsize(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
a integer
}
\description{
Returns the program size, a very approximate measure of a regexp's "cost".
Larger numbers are more expensive than smaller numbers.
}
\examples{
regexp = re2_compile("1")
get_programsize(regexp)
}

