% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplify.R
\name{get_simplify}
\alias{get_simplify}
\title{Simplify pattern strings.}
\usage{
get_simplify(pattern, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{...}{further arguments to compile input character.}

\item{input}{a character vector}
}
\description{
Simplify pattern strings.
}
\examples{
get_simplify("a{1}")
}

