% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%=~\%}
\alias{\%!~\%}
\alias{\%=~\%}
\title{Check strings with a regular expression}
\usage{
string \%=~\% pattern

string \%!~\% pattern
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}
}
\description{
return re2_match(pattern, string)
}
\examples{
c("pt","sd") \%=~\% "sd"
c("pt","sd") \%!~\% "sd"
}

