% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%<~\%}
\alias{\%<~\%}
\title{Create a pre-compiled regular expression}
\usage{
varible \%<~\% pattern
}
\arguments{
\item{varible}{varible name}

\item{pattern}{regular expression pattern}
}
\description{
Create a pre-compiled regular expression
}
\examples{
regexp \%<~\% "(12)"
regexp
}
\seealso{
more options with \code{\link{re2}}
}

