% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_re2c_na}
\alias{is_re2c_na}
\title{Check NA pattern}
\usage{
is_re2c_na(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
boolean
}
\description{
Returns whether a pre-compiled regular expression is NA.
}
\examples{
is_re2c_na(re2(NA))
}

