% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.re2_matrix}
\alias{print.re2_matrix}
\title{Print information about a re2_matrix}
\usage{
\method{print}{re2_matrix}(x, ...)
}
\arguments{
\item{x}{an re2_matrix}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print information about a re2_matrix
}
\examples{
print(re2_match(".*", "abc"))
}
