% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.re2c}
\alias{print.re2c}
\title{Print information about a pre-compiled regular expression}
\usage{
\method{print}{re2c}(x, options = FALSE, ...)
}
\arguments{
\item{x}{a pre-compiled regular expression}

\item{options}{print options for the pre-compiled regular expression}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print information about a pre-compiled regular expression
}
\examples{
re2("(.*)@([^.]*)")
re2("(?P<name>sd)")
}

