% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{re2}
\alias{re2}
\title{Create a pre-compiled regular expression}
\usage{
re2(pattern, utf_8 = TRUE, case_sensitive = TRUE, posix_syntax = FALSE,
  dot_nl = FALSE, literal = FALSE, longest_match = FALSE,
  never_nl = FALSE, never_capture = FALSE, one_line = FALSE,
  perl_classes = FALSE, word_boundary = FALSE, log_error = FALSE,
  max_mem = 8388608, simplify = TRUE)
}
\arguments{
\item{pattern}{regular expression pattern}

\item{utf_8}{(true)  text and pattern are UTF-8; otherwise Latin-1}

\item{case_sensitive}{(true)  match is case-sensitive (regexp can override with (?i) unless in posix_syntax mode)}

\item{posix_syntax}{(false) restrict regexps to POSIX egrep syntax}

\item{dot_nl}{(false) dot matches everything including new line}

\item{literal}{(false) interpret string as literal, not regexp}

\item{longest_match}{(false) search for longest match, not first match}

\item{never_nl}{(false) never match \\n, even if it is in regexp}

\item{never_capture}{(false) parse all parens as non-capturing}

\item{one_line}{(false) ^ and $ only match beginning and end of text, when posix_syntax == false this features are always enabled}

\item{perl_classes}{(false) allow Perl's \\d \\s \\w \\D \\S \\W, when posix_syntax == false this features are always enabled}

\item{word_boundary}{(false) allow Perl's \\b \\B (word boundary and not), when posix_syntax == false this features are always enabled}

\item{log_error}{(false) log syntax and execution errors}

\item{max_mem}{(see details)  approx. max memory footprint of RE2}

\item{simplify}{(true) return a object instead of a list when pattern length is 1.}
}
\value{
a pre-compiled regular expression
}
\description{
Create a pre-compiled regular expression from a string.
}
\details{
The max_mem option controls how much memory can be used
to hold the compiled form of the regexp (the Prog) and
its cached DFA graphs.  Code Search placed limits on the number
of Prog instructions and DFA states: 10,000 for both.
In RE2, those limits would translate to about 240 KB per Prog
and perhaps 2.5 MB per DFA (DFA state sizes vary by regexp;

RE2 does a better job of keeping them small than Code Search did).
Each RE2 has two Progs (one forward, one reverse), and each Prog
can have two DFAs (one first match, one longest match).
That makes 4 DFAs:

    forward, first-match    - used for UNANCHORED or ANCHOR_LEFT searches
if opt.longest_match() == false

forward, longest-match  - used for all ANCHOR_BOTH searches,
and the other two kinds if
opt.longest_match() == true

reverse, first-match    - never used

reverse, longest-match  - used as second phase for unanchored searches

The RE2 memory budget is statically divided between the two
Progs and then the DFAs: two thirds to the forward Prog
and one third to the reverse Prog.  The forward Prog gives half
of what it has left over to each of its DFAs.  The reverse Prog
gives it all to its longest-match DFA.

Once a DFA fills its budget, it flushes its cache and starts over.
If this happens too often, RE2 falls back on the NFA implementation.

For now, make the default budget something close to Code Search.

DefaultMaxMem = 8<<20;
}
\examples{
regexp = re2("test")
regexp
}

