% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.R
\name{re2_detect}
\alias{re2_detect}
\title{Test a pattern in strings, and return boolean.}
\usage{
re2_detect(string, pattern, anchor = UNANCHORED, parallel = FALSE,
  grain_size = 1e+05, ...)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{UNANCHORED: no anchor. ANCHOR_START: anchor match at the beginning of the string. ANCHOR_BOTH: anchor match at the beginning and the end of the string.}

\item{parallel}{use multithread}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Test a pattern in strings, and return boolean.
}
\examples{
re2_detect("one", "(o.e)")
re2_detect("123-234-2222", "\\\\d\\\\d\\\\d-\\\\d\\\\d\\\\d-\\\\d\\\\d\\\\d\\\\d")

words = c("sunny","beach","happy","really")
re2_detect(words, "y")
re2_detect(words, "^b")
re2_detect(words, "[abc]")

# vectorize
(res = re2_detect("This", letters))
letters[res]

}

