% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{re2_extract}
\alias{re2_extract}
\alias{re2_extract_all}
\title{Extract a pattern in strings.}
\usage{
re2_extract(input, pattern, parallel = FALSE, grain_size = 1e+05, ...)

re2_extract_all(input, pattern, parallel = FALSE, grain_size = 1e+05, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{parallel}{use multithread}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Like re2_replace, except that if the pattern matches, substring is copied into the result. The non-matching
portions of "text" are ignored.
}
\examples{
re2_extract("yabba dabba doo", "(.)")
}

