% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{re2_fullmatch}
\alias{re2_fullmatch}
\title{re2_fullmatch}
\usage{
re2_fullmatch(pattern, string, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{input}{a character vector}
}
\description{
re2_fullmatch
}

