% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_match}
\alias{re2_match}
\alias{re2_match_all}
\title{Find matched groups from strings.}
\usage{
re2_match(string, pattern, anchor = UNANCHORED, parallel = FALSE,
  grain_size = 1e+05, ...)

re2_match_all(string, pattern, anchor = UNANCHORED, parallel = FALSE,
  grain_size = 1e+05, ...)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{UNANCHORED: no anchor. ANCHOR_START: anchor match at the beginning of the string. ANCHOR_BOTH: anchor match at the beginning and the end of the string.}

\item{parallel}{multithreading support}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Find matched groups from strings.
}
\examples{

test_string = "this is just one test";
re2_match(test_string, "(o.e)")

(res = re2_match(test_string, "(o.e)"))
str(res)

(res = re2_match(test_string, "(?P<testname>this)( is)"))
str(res)

(res = re2_match_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test");
(res = re2_match_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test")
(res = re2_match(test_string, "is"))

regexp = re2("test",case_sensitive = FALSE)
re2_match("TEST", regexp)

}

