% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{re2_pextract}
\alias{re2_pextract}
\alias{re2_pextract_all}
\title{Extract a pattern in strings with multithread.}
\usage{
re2_pextract(input, pattern, all = FALSE, ...)

re2_pextract_all(input, pattern, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{all}{extract all patterns}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Like re2_preplace, except that if the pattern matches, substring is copied into the result. The non-matching
portions of "text" are ignored.
}
\examples{
re2_pextract("yabba dabba doo", "(.)")
}

