% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{re2_plocate}
\alias{re2_plocate}
\title{Locate a pattern in strings with multithread.}
\usage{
re2_plocate(input, pattern, grain_size = 1e+05, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Like re2_preplace, except that if the pattern matches, substring is copied into the result. The non-matching
portions of "text" are ignored.
}
\examples{
re2_plocate("yabba dabba doo", "(d)")
}

