% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{re2_psubset}
\alias{re2_psubset}
\title{Keep strings matching a pattern with multithread.}
\usage{
re2_psubset(input, pattern, anchor = 0, grain_size = 1e+05, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{a positive number. 0: no anchor. 1: anchor match at the beginning of the string. 2 or larger number: anchor match at the beginning and the end of the string.}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A character vector.
}
\description{
This is a convenient wrapper around \code{x[re2_detect(x, pattern)]}.
}

