% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{re2_replace_all}
\alias{re2_replace_all}
\alias{re2_replace_all.character}
\alias{re2_replace_all.re2exp}
\title{Replaces successive non-overlapping occurrences}
\usage{
re2_replace_all(pattern, rewrite, input, ...)

\method{re2_replace_all}{re2exp}(pattern, rewrite, input, ...)

\method{re2_replace_all}{character}(pattern, rewrite, input, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{rewrite}{replace the first match of "pattern" in "input" with "rewrite"}

\item{input}{a character vector}
}
\value{
a character vector
}
\description{
Like re2_replace, except replaces successive non-overlapping occurrences
of the pattern in the string with the rewrite.
}
\details{
Within "rewrite", backslash-escaped digits (\\1 to \\9) can be
used to insert text matching corresponding parenthesized group
from the pattern.  \\0 in "rewrite" refers to the entire matching
text.
}
\examples{
regexp = re2_compile("b+")
re2_replace_all(regexp,"d", "yabba dabba doo") == "yada dada doo"
re2_replace_all("b+","d", "yabba dabba doo") == "yada dada doo"
}

