% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{re2_test_match}
\alias{re2_test_match}
\title{re2_test_match}
\usage{
re2_test_match(pattern, string, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{input}{a character vector}
}
\description{
re2_test_match
}

