% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{show_regex}
\alias{show_regex}
\title{Show regex pattern in a htmlwidget}
\usage{
show_regex(pattern, width = NULL, height = NULL)
}
\arguments{
\item{pattern}{a pattern string}

\item{width}{the widget width}

\item{height}{the widget height}
}
\description{
Show regex pattern in a htmlwidget.
}

