% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_expression_size}
\alias{get_expression_size}
\title{Get pre-compiled regular expression program size}
\usage{
get_expression_size(ptr)
}
\arguments{
\item{ptr}{a pre-compiled regular expression}
}
\value{
a integer
}
\description{
Returns the program size, a very approximate measure of a regexp's "cost".
Larger numbers are more expensive than smaller numbers.
}
\examples{
regexp = re2("1")
get_expression_size(regexp)
}

