% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_options}
\alias{get_options}
\title{Get options of a pre-compiled regular expression}
\usage{
get_options(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
an R list
}
\description{
Returns options of a pre-compiled regular expression
}
\examples{
get_options(re2("test"))
}

