% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/compile.R
\name{cpp_get_pattern}
\alias{cpp_get_pattern}
\alias{get_pattern}
\title{The string specification for this RE2.}
\usage{
cpp_get_pattern(regexp)

get_pattern(x)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
a string
}
\description{
The string specification for this RE2.
}
\examples{
regexp = re2_compile("1")
get_pattern(regexp)
}

