% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{get_pattern}
\alias{get_pattern}
\title{The string specification for this RE2.}
\usage{
get_pattern(regexp)
}
\arguments{
\item{regexp}{a pre-compiled regular expression}
}
\value{
a string
}
\description{
The string specification for this RE2.
}
\examples{
regexp = re2("1")
get_pattern(regexp)
}

