% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{get_program_fanout}
\alias{get_program_fanout}
\title{Get program fanout}
\usage{
get_program_fanout(pattern, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{...}{further arguments passed to \code{\link{re2}}}
}
\description{
Return the program fanout as a histogram bucketed by powers of 2.
}
\examples{
re1 = re2("(?:(?:(?:(?:(?:.)?){1})*)+)")
re10 = re2("(?:(?:(?:(?:(?:.)?){10})*)+)")
re100 = re2("(?:(?:(?:(?:(?:.)?){100})*)+)")
re1000 = re2("(?:(?:(?:(?:(?:.)?){1000})*)+)")

get_program_fanout(re1)
get_program_fanout(re10)
get_program_fanout(re100)
get_program_fanout(re1000)
}
