% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{\%=~\%}
\alias{\%=~\%}
\title{check a regular expression}
\usage{
pattern \%=~\% string
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{string}{a character vector}
}
\description{
return re2_match(pattern, string)
}
\examples{
"sd" \%=~\% c("pt","sd")
"sd" \%!~\% c("pt","sd")
}

