% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{is_re2c_na}
\alias{is_re2c_na}
\title{Check NA pattern}
\usage{
is_re2c_na(pattern, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{...}{further arguments passed to \code{\link{re2}}}
}
\value{
a boolean
}
\description{
Returns whether a pre-compiled regular expression is NA.
}
\examples{
is_re2c_na(re2(NA))
}

