% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{quote_meta}
\alias{quote_meta}
\title{Escapes all potentially meaningful regexp characters in  'unquoted'.}
\usage{
quote_meta(unquoted, parallel = FALSE)
}
\arguments{
\item{unquoted}{unquoted string}

\item{parallel}{multithreading support}
}
\value{
quoted string
}
\description{
The returned string, used as a regular expression, will exactly match the original string.  For example,
}
\details{
1.5-2.0

may become:

1\\.5\-2\\.0
}
\examples{
quote_meta(c("1.2","abc"))
}

