% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{re2_extract}
\alias{re2_extract}
\alias{re2_extract.character}
\alias{re2_extract.re2exp}
\title{Extract one matched patterns in a string.}
\usage{
re2_extract(pattern, input, rewrite = "\\\\1", ...)

\method{re2_extract}{re2exp}(pattern, input, rewrite = "\\\\1", ...)

\method{re2_extract}{character}(pattern, input, rewrite = "\\\\1", ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{input}{a character vector}

\item{rewrite}{replace the first match of "pattern" in "input" with "rewrite"}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Like Replace, except that if the pattern matches, "rewrite"
is copied into "out" with substitutions.  The non-matching
portions of "text" are ignored.
}
\details{
Returns true iff a match occurred and the extraction happened
}
\examples{
re2_extract("(.)","yabba dabba doo")
re2_extract("(.*)@([^.]*)","test@me.com","\\\\2!\\\\1")
}

