% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{re2_extract}
\alias{re2_extract}
\alias{re2_extract.character}
\alias{re2_extract.re2exp}
\title{Extract matched patterns in a string.}
\usage{
re2_extract(pattern, rewrite, input, ...)

\method{re2_extract}{re2exp}(pattern, rewrite, input)

\method{re2_extract}{character}(pattern, rewrite, input, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{rewrite}{replace the first match of "pattern" in "input" with "rewrite"}

\item{input}{a character vector}
}
\description{
Like Replace, except that if the pattern matches, "rewrite"
is copied into "out" with substitutions.  The non-matching
portions of "text" are ignored.
}
\details{
Returns true iff a match occurred and the extraction happened
}

