% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{re2_locate}
\alias{re2_locate}
\alias{re2_locate_all}
\title{Locate a pattern in strings.}
\usage{
re2_locate(string, pattern, parallel = FALSE, grain_size = 1e+05, ...)

re2_locate_all(string, pattern, parallel = FALSE, grain_size = 1e+05, ...)
}
\arguments{
\item{string}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{parallel}{use multithread}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms}

\item{...}{further arguments passed to or from other methods.}
}
\description{
If the match is of length 0, (e.g. from a special match like $) end will be one character less than start.
}
\examples{
re2_locate("yabba dabba doo", "d")
}

