% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_match}
\alias{re2_match}
\alias{re2_match.character}
\alias{re2_match.re2exp}
\title{Match patterns in a string.}
\usage{
re2_match(pattern, string, value = FALSE, anchor = "none", all = FALSE,
  ...)

\method{re2_match}{re2exp}(pattern, string, value = FALSE, anchor = "none",
  all = FALSE, ...)

\method{re2_match}{character}(pattern, string, value = FALSE,
  anchor = "none", all = FALSE, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{string}{a character vector}

\item{value}{return value instead of bool result}

\item{anchor}{"start": anchor match at the beginning of the string, "both": anchor match at the beginning and the end of the string, "none": no anchor.}

\item{all}{find all matches instead of the first match. When result = "value", a matched character matrix will be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Match patterns in a string.
}

