% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_match}
\alias{re2_match}
\alias{re2_match_all}
\title{Find matched groups from strings.}
\usage{
re2_match(input, pattern, anchor = 0, value = TRUE, all = FALSE,
  parallel = FALSE, ...)

re2_match_all(input, pattern, anchor = 0, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{a positive number. 0: no anchor. 1: anchor match at the beginning of the string. 2 or larger number: anchor match at the beginning and the end of the string.}

\item{value}{return value instead of bool result}

\item{all}{find all matches instead of the first match. When value is true, a matched character matrix or a list will be returned.}

\item{parallel}{multithreading support}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Find matched groups from strings.
}
\examples{

test_string = "this is just one test";
re2_match(test_string, "(o.e)")

(res = re2_match(test_string, "(o.e)"))
str(res)

(res = re2_match(test_string, "(?P<testname>this)( is)"))
str(res)

(res = re2_match_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test");
(res = re2_match_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test")
(res = re2_match(test_string, "is"))

regexp = re2("test",case_sensitive = FALSE)
re2_match("TEST", regexp)

}

