% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R, R/search.R
\name{\%=~\%}
\alias{\%!~\%}
\alias{\%=~\%}
\alias{re2_match}
\alias{re2_match.character}
\alias{re2_match.re2exp}
\title{re2_match}
\usage{
pattern \%=~\% string

pattern \%!~\% string

re2_match(pattern, string, result = "bool", anchor = "start", all = TRUE,
  ...)

\method{re2_match}{re2exp}(pattern, string, result = "bool",
  anchor = "start", all = TRUE, ...)

\method{re2_match}{character}(pattern, string, result = "bool",
  anchor = "start", all = TRUE, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{string}{a character vector}

\item{result}{"bool", "value", "object"}

\item{anchor}{"start": anchor match at the beginning of the string, "both": anchor match at the beginning and the end of the string, "none": no anchor.}

\item{all}{find all match instead of first match. When all = FALSE and result = "value", the return type will be character vector when there is no capture group or data.frame.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
re2_match
}

