% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R, R/search.R
\name{\%=~\%}
\alias{\%!~\%}
\alias{\%=~\%}
\alias{re2_match}
\alias{re2_match.character}
\alias{re2_match.re2exp}
\title{re2_match}
\usage{
pattern \%=~\% string

pattern \%!~\% string

re2_match(pattern, string, result, anchor, ...)

\method{re2_match}{re2exp}(pattern, string, result, anchor, ...)

\method{re2_match}{character}(pattern, string, result, anchor, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{string}{a character vector}

\item{result}{bool, value or object}

\item{anchor}{"match": anchor match at the beginning of the string, "fullmatch": anchor match at the beginning and the end of the string, "search": no anchor.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
re2_match
}

