% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{re2_match}
\alias{re2_match}
\alias{re2_match.character}
\alias{re2_match.re2exp}
\title{re2_match}
\usage{
re2_match(pattern, string, value = FALSE, anchor = "none", all = TRUE,
  ...)

\method{re2_match}{re2exp}(pattern, string, value = FALSE, anchor = "none",
  all = TRUE, ...)

\method{re2_match}{character}(pattern, string, value = FALSE,
  anchor = "none", all = TRUE, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{string}{a character vector}

\item{value}{return value instead of bool result}

\item{anchor}{"start": anchor match at the beginning of the string, "both": anchor match at the beginning and the end of the string, "none": no anchor.}

\item{all}{find all match instead of first match. When all = FALSE and result = "value", the return type will be character vector when there is no capture group or data.frame.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
re2_match
}

