% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_match}
\alias{re2_match}
\title{Match patterns in a string.}
\usage{
re2_match(input, pattern, value = FALSE, anchor = "none", all = FALSE,
  ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{value}{return value instead of bool result}

\item{anchor}{"start": anchor match at the beginning of the string, "both": anchor match at the beginning and the end of the string, "none": no anchor.}

\item{all}{find all matches instead of the first match. When result = "value", a matched character matrix will be returned.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Match patterns in a string.
}
\examples{

test_string = "this is just one test";
re2_match(test_string, "(o.e)")

(res = re2_match(test_string, "(o.e)", value = TRUE))
str(res)

(res = re2_match(test_string, "(?P<testname>this)( is)", value = TRUE))
str(res)

(res = re2_match(test_string, "(is)", value = TRUE, all = TRUE))

test_string = c("this is just one test", "the second test");
(res = re2_match(test_string, "(is)", value = TRUE, all = TRUE))

test_string = c("this is just one test", "the second test")
(res = re2_match(test_string, "is", value = TRUE))

regexp = re2("test",case_sensitive = FALSE)
re2_match("TEST", regexp)
}

