% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_pdetect}
\alias{re2_pdetect}
\title{Test a pattern in strings with multithread, and return boolean.}
\usage{
re2_pdetect(input, pattern, anchor = 0, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{a positive number. 0: no anchor. 1: anchor match at the beginning of the string. 2 or larger number: anchor match at the beginning and the end of the string.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Test a pattern in strings with multithread, and return boolean.
}
\examples{
re2_pdetect(test_string, "(o.e)")
}

