% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{re2_pmatch}
\alias{re2_pmatch}
\alias{re2_pmatch_all}
\title{Find matched groups from strings with multithread.}
\usage{
re2_pmatch(input, pattern, anchor = 0, grain_size = 1e+05, ...)

re2_pmatch_all(input, pattern, anchor = 0, grain_size = 1e+05, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{a positive number. 0: no anchor. 1: anchor match at the beginning of the string. 2 or larger number: anchor match at the beginning and the end of the string.}

\item{grain_size}{a minimum chunk size for tuning the behavior of parallel algorithms.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Find matched groups from strings with multithread.
}
\examples{

test_string = "this is just one test";
re2_pdetect(test_string, "(o.e)")

(res = re2_pmatch(test_string, "(o.e)"))
str(res)

(res = re2_pmatch(test_string, "(?P<testname>this)( is)"))
str(res)

(res = re2_pmatch_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test");
(res = re2_pmatch_all(test_string, "(is)"))

test_string = c("this is just one test", "the second test")
(res = re2_pmatch_all(test_string, "is"))

regexp = re2("test",case_sensitive = FALSE)
re2_pmatch("TEST", regexp)

}

