% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{re2_psplit}
\alias{re2_psplit}
\alias{re2_psplit_fixed}
\title{Split strings with a pattern with multithread.}
\usage{
re2_psplit(input, pattern, part = Inf, ...)

re2_psplit_fixed(input, pattern, part, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{part}{number of pieces to return. Default (Inf) uses all possible split positions.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Split up a string into pieces.
}
\examples{
re2_psplit("yabba dabba doo", " ")
}

