% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{re2_replace}
\alias{re2_replace}
\alias{re2_replace_all}
\title{Replace a pattern in strings.}
\usage{
re2_replace(input, pattern, rewrite, ...)

re2_replace_all(input, pattern, rewrite, ...)
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{rewrite}{replace the first match or all of the match of "pattern" in "input" with "rewrite"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a character vector
}
\description{
Replace the first match of "pattern" in "str" with "rewrite".
Within "rewrite", backslash-escaped digits (\\1 to \\9) can be
used to insert text matching corresponding parenthesized group
from the pattern.  \\0 in "rewrite" refers to the entire matching
text.
}
\examples{
regexp = re2("b+")
re2_replace_all("yabba dabba doo", regexp,"d") == "yada dada doo"
re2_replace("yabba dabba doo", "b+","d") == "yada dabba doo"
}

