% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace.R
\name{re2_replace}
\alias{re2_replace}
\alias{re2_replace.character}
\alias{re2_replace.re2exp}
\title{Replace matched patterns in a string.}
\usage{
re2_replace(pattern, rewrite, input, global = FALSE, ...)

\method{re2_replace}{re2exp}(pattern, rewrite, input, global = FALSE, ...)

\method{re2_replace}{character}(pattern, rewrite, input, global = FALSE, ...)
}
\arguments{
\item{pattern}{a pre-compiled regular expression or a string}

\item{rewrite}{replace the first match or all of the match of "pattern" in "input" with "rewrite"}

\item{input}{a character vector}

\item{global}{if it is TRUE, it will replaces successive non-overlapping occurrences}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a character vector
}
\description{
Replace the first match of "pattern" in "str" with "rewrite".
Within "rewrite", backslash-escaped digits (\\1 to \\9) can be
used to insert text matching corresponding parenthesized group
from the pattern.  \\0 in "rewrite" refers to the entire matching
text.
}
\examples{
regexp = re2("b+")
re2_replace(regexp,"d", "yabba dabba doo") == "yada dada doo"
re2_replace("b+","d", "yabba dabba doo", global = FALSE) == "yada dabba doo"
}

