% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{re2_subset}
\alias{re2_subset}
\alias{re2_subset<-}
\title{Keep strings matching a pattern.}
\usage{
re2_subset(input, pattern, anchor = 0, ...)

re2_subset(input, pattern, anchor = 0, ...) <- value
}
\arguments{
\item{input}{a character vector}

\item{pattern}{a pre-compiled regular expression or a string}

\item{anchor}{a positive number. 0: no anchor. 1: anchor match at the beginning of the string. 2 or larger number: anchor match at the beginning and the end of the string.}

\item{...}{further arguments passed to or from other methods.}

\item{value}{character vector to be substituted with; replacement function only}
}
\value{
A character vector.
}
\description{
This is a convenient wrapper around \code{x[re2_detect(x, pattern)]}.
}
\examples{
fruit <- c("apple", "banana", "pear", "pinapple")
re2_subset(fruit, "a")
re2_subset(fruit, "^a")
re2_subset(fruit, "a$")
re2_subset(fruit, "b")
re2_subset(fruit, "[aeiou]")

re2_subset(c("a", NA, "b"), ".")
}

