% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widget.R
\name{show_regex}
\alias{show_regex}
\title{Show regex pattern in a htmlwidget}
\usage{
show_regex(pattern, width = NULL, height = NULL)
}
\arguments{
\item{pattern}{a pattern string}

\item{width}{the widget width}

\item{height}{the widget height}
}
\description{
Show JS-style regex pattern in an htmlwidget.
}
\details{
Most parts of RE2 regex syntax are supported, except for some special Unicode character classes.
}
\examples{
# US ZIP code

show_regex("[0-9]{5}(?:-[0-9]{4})?")

# Email

show_regex("\\\\b[a-zA-Z0-9._\%-]+@[a-zA-Z0-9.-]+\\\\.[a-zA-Z]{2,4}\\\\b")

# Hex value

show_regex("#?([a-f0-9]{6}|[a-f0-9]{3})")

# URL

show_regex("((www\\\\.|(http|https|ftp|news|file)+\\\\:\\\\/\\\\/)[_.a-z0-9-]+\\\\.[a-z0-9\\\\/_:@=.+?,##\%&~-]*[^.|\\\\'|\\\\# |!|\\\\(|?|,| |>|<|;|\\\\)])")

# major credit cards

show_regex("^(?:4[0-9]{12}(?:[0-9]{3})?|5[1-5][0-9]{14}|6011[0-9]{12}|622((12[6-9]|1[3-9][0-9])|([2-8][0-9][0-9])|(9(([0-1][0-9])|(2[0-5]))))[0-9]{10}|64[4-9][0-9]{13}|65[0-9]{14}|3(?:0[0-5]|[68][0-9])[0-9]{11}|3[47][0-9]{13})*$")

}
\seealso{
https://regexper.com/documentation.html
}

